Refactor: Rename ProjectSidebar to ProjectInfoPanel
Renamed ProjectSidebar and ProjectSidebarSkeleton to ProjectInfoPanel and ProjectInfoPanelSkeleton respectively to more accurately reflect their functionality. Updated all relevant imports and usages.
