+++ src/renderer/components/skeleton/index.ts
@@ -2,4 +2,4 @@ export { default as EditorViewSkeleton } from './EditorViewSkeleton';
 export { default as SettingsViewSkeleton } from './SettingsViewSkeleton';
 export { default as AboutViewSkeleton } from './AboutViewSkeleton';
 export { default as ProjectManagerSkeleton } from './ProjectManagerSkeleton';
-export { default as ProjectSidebarSkeleton } from './ProjectSidebarSkeleton';
+export { default as ProjectInfoPanelSkeleton } from './ProjectInfoPanelSkeleton';
